fix: activeEditor function is now async with retry logic
This is a temporary work-around until there is a way to be notified of the very first document being opened.

The activeEditor function will throw if it cannot get an activeWindow after 5 times. Throw is appropriate here because if the extension was activated without an activeWindow in the space of 2.5 seconds (5 * 500milliseconds), then the current state is not good.

A nice benefit from this code is that the `if(!editor)` are now not needed.